+++ src/Elders.Cronus.Persistence.Cassandra/CassandraEventStore.cs
@@ -21,14 +21,20 @@ namespace Elders.Cronus.Persistence.Cassandra
 
         private readonly ICassandraEventStoreTableNameStrategy tableNameStrategy;
 
+        private readonly ConsistencyLevel writeConsistencyLevel;
+
+        private readonly ConsistencyLevel readConsistencyLevel;
+
         private readonly ConcurrentDictionary<string, PreparedStatement> persistAggregateEventsPreparedStatements;
         private readonly ConcurrentDictionary<string, PreparedStatement> loadAggregateEventsPreparedStatements;
 
-        public CassandraEventStore(ISession session, ICassandraEventStoreTableNameStrategy tableNameStrategy, ISerializer serializer)
+        public CassandraEventStore(ISession session, ICassandraEventStoreTableNameStrategy tableNameStrategy, ISerializer serializer, ConsistencyLevel writeConsistencyLevel, ConsistencyLevel readConsistencyLevel)
         {
             this.tableNameStrategy = tableNameStrategy;
             this.session = session;
             this.serializer = serializer;
+            this.writeConsistencyLevel = writeConsistencyLevel;
+            this.readConsistencyLevel = readConsistencyLevel;
             this.persistAggregateEventsPreparedStatements = new ConcurrentDictionary<string, PreparedStatement>();
             this.loadAggregateEventsPreparedStatements = new ConcurrentDictionary<string, PreparedStatement>();
         }
@@ -36,12 +42,14 @@ namespace Elders.Cronus.Persistence.Cassandra
         private PreparedStatement GetPreparedStatementToPersistAnAggregateCommit(AggregateCommit aggregateCommit)
         {
             PreparedStatement persistAggregatePreparedStatement;
-            if (!persistAggregateEventsPreparedStatements.TryGetValue(aggregateCommit.BoundedContext, out persistAggregatePreparedStatement))
+            if (persistAggregateEventsPreparedStatements.TryGetValue(aggregateCommit.BoundedContext, out persistAggregatePreparedStatement) == false)
             {
                 string tableName = tableNameStrategy.GetEventsTableName(aggregateCommit);
                 persistAggregatePreparedStatement = session.Prepare(String.Format(InsertEventsQueryTemplate, tableName));
                 persistAggregateEventsPreparedStatements.TryAdd(aggregateCommit.BoundedContext, persistAggregatePreparedStatement);
             }
+
+            persistAggregatePreparedStatement.SetConsistencyLevel(writeConsistencyLevel);
 
             return persistAggregatePreparedStatement;
         }
@@ -83,12 +91,14 @@ namespace Elders.Cronus.Persistence.Cassandra
         private PreparedStatement GetPreparedStatementToLoadAnAggregateCommit(string boundedContext)
         {
             PreparedStatement loadAggregatePreparedStatement;
-            if (!loadAggregateEventsPreparedStatements.TryGetValue(boundedContext, out loadAggregatePreparedStatement))
+            if (loadAggregateEventsPreparedStatements.TryGetValue(boundedContext, out loadAggregatePreparedStatement) == false)
             {
                 string tableName = tableNameStrategy.GetEventsTableName(boundedContext);
                 loadAggregatePreparedStatement = session.Prepare(String.Format(LoadAggregateEventsQueryTemplate, tableName));
                 loadAggregateEventsPreparedStatements.TryAdd(boundedContext, loadAggregatePreparedStatement);
             }
+
+            loadAggregatePreparedStatement.SetConsistencyLevel(readConsistencyLevel);
 
             return loadAggregatePreparedStatement;
         }